refactor(kubernetes): move react codebase [EE-3349]
closes [EE-3349]
