+++ app/kubernetes/react/views/index.ts
@@ -1,11 +1,11 @@
 import angular from 'angular';
 
 import { r2a } from '@/react-tools/react2angular';
-import { IngressesDatatableView } from '@/kubernetes/react/views/networks/ingresses/IngressDatatable';
-import { CreateIngressView } from '@/kubernetes/react/views/networks/ingresses/CreateIngressView';
 import { withCurrentUser } from '@/react-tools/withCurrentUser';
 import { withReactQuery } from '@/react-tools/withReactQuery';
 import { withUIRouter } from '@/react-tools/withUIRouter';
+import { IngressesDatatableView } from '@/react/kubernetes/ingresses/IngressDatatable';
+import { CreateIngressView } from '@/react/kubernetes/ingresses/CreateIngressView';
 
 export const viewsModule = angular
   .module('portainer.kubernetes.react.views', [])
