+++ app/react/kubernetes/ingresses/IngressDatatable/columns/className.tsx
