+++ app/react/kubernetes/ingresses/IngressDatatable/columns/namespace.tsx
