+++ app/react/kubernetes/ingresses/IngressDatatable/columns/type.tsx
