+++ app/react/kubernetes/ingresses/style.css
