+++ app/react/kubernetes/ingresses/types.ts
