+++ app/react/kubernetes/networks/services/service.ts
