+++ app/react/kubernetes/services/types/v1IngressClass.ts
